server/requirements.txt
flask~=1.1.2
flask-sqlalchemy~=2.4.1
flask-migrate~=2.5.2
flask-script~=2.0.6
psycopg2~=2.8.4